+++ ss1/content/movie/compression/ControlWord.go
@@ -20,7 +20,7 @@ func (word ControlWord) IsLongOffset() bool {
 
 // LongOffset returns the long offset value. Only relevant if IsLongOffset() returns true.
 func (word ControlWord) LongOffset() uint32 {
-	return uint32((uint32(word) >> 0) & 0xFFFFF)
+	return (uint32(word) >> 0) & 0xFFFFF
 }
 
 // Type returns the type of the control. Only relevant if IsLongOffset() returns false.
@@ -30,5 +30,5 @@ func (word ControlWord) Type() ControlType {
 
 // Parameter returns the type specific parameter value.
 func (word ControlWord) Parameter() uint32 {
-	return uint32((uint32(word) >> 0) & 0x1FFFF)
+	return (uint32(word) >> 0) & 0x1FFFF
 }
